+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.2.0] - 2022-03-21
+
+### Added
+
+-   Add `plugin:install` console command to install WordPress plugins via composer
+-   Add `plugin:uninstall` console command to uninstall WordPress plugins via composer
+-   Add `plugin:update` console command to update WordPress plugins via composer
+-   Add `CHANGELOG.md` file
+
+### Changed
+
+-   Update `README.md` **Console commands** section
+
 ## [1.1.0] - 2021-07-27
 
 ### Added
@@ -9,3 +22,7 @@
 
 -   `theme:update` console command source code improvements
 -   Update `README.md` **Console commands** section
+
+## [1.0.0] - 2021-07-19
+
+-   Initial release